+++ src/automated-polymorpher.ts
@@ -119,10 +119,9 @@ Hooks.once('libChangelogsReady', function () {
   libChangelogs.register(
     CONSTANTS.MODULE_NAME,
     `
-  - Add hud control
-  - add some new setting
-  - Add a fast polymorphing mechanism for make the comabt more fluid
-  - Add socketLib, API, new feature
+    - Add check for show the hud button only if at least a polymorphing actor is present on the current actor
+    - Converted the hud settings from 'world' to 'client'
+    - Set module settings 'hudEnable' default to true
   `,
     'minor',
   );