+++ changelog.md
@@ -1,3 +1,9 @@
+## 1.0.17
+
+- Add check for show the hud button only if at least a polymorphing actor is present on the current actor
+- Converted the hud settings from 'world' to 'client'
+- Set module settings 'hudEnable' default to true
+
 ## 1.0.16
 
 - Add hud control